event_management_service/requirements.txt
flask 
grpcio 
grpcio-tools 
pymysql
mysql-connector-python
flask-cors